+++ lib/SecondTap.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class SecondTab extends StatelessWidget {
+  const SecondTab({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Text(
+          'First Tab',
+          style: TextStyle(fontSize: 24),
+        ),
+      ),
+    );
+  }
+}
